+++ eslint.config.mjs
@@ -32,7 +32,11 @@ export default tseslint.config(
       }
     },
     rules: {
-      "@typescript-eslint/no-floating-promises": "error"
+      "@typescript-eslint/no-floating-promises": "error",
+      "@typescript-eslint/no-unused-vars": [
+        "error",
+        { argsIgnorePattern: "^_", varsIgnorePattern: "^_" }
+      ]
     }
   }
 );